+++ src/propulsion/can/can_handler.hpp
@@ -1,6 +1,20 @@
-//
-// Created by Gregor Konzett on 2019-01-28.
-//
+/*
+ * Author: Gregor Konzett
+ * Organisation: HYPED
+ * Date: 
+ * Description:
+ *
+ *    Copyright 2019 HYPED
+ *    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *    except in compliance with the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software distributed under
+ *    the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ *    either express or implied. See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
 
 #ifndef HYPED_2019_CANHANDLER_HPP
 #define HYPED_2019_CANHANDLER_HPP
@@ -12,7 +26,7 @@
 #include "utils/io/can.hpp"
 #include "utils/logger.hpp"
 #include "can_sender.hpp"
-#include "propulsion/sender_interface.hpp"
+#include "propulsion/can/sender_interface.hpp"
 
 namespace hyped {
